NOT WORKING: pre-commit for H2 database; added parameters parsing, via reflection and annotation

+++ src/main/java/sfgamedataeditor/fieldwrapping/Data.java
@@ -5,7 +5,7 @@ import java.lang.annotation.RetentionPolicy;
 
 @Retention(RetentionPolicy.RUNTIME)
 public @interface Data {
-    long offset();
+    int offset();
 
     int length();
 }
